chore(views): move views to a dedicated file
Also, setup a dedicated egapro command for `serve` (vs. makefile)
