+++ Makefile
@@ -3,12 +3,6 @@
 develop:
 	pip install -e .[dev,test]
 
-serve: SHELL := python3
-serve:
-	from egapro import serve
-	serve(reload=True)
-.PHONY: serve
-
 init: SHELL := python3
 init:
 	import asyncio
